+++ packages/astro/src/core/errors.ts
@@ -9,6 +9,7 @@ export interface ErrorWithMetadata {
 	[name: string]: any;
 	message: string;
 	stack: string;
+	hint?: string;
 	id?: string;
 	frame?: string;
 	plugin?: string;
@@ -38,6 +39,24 @@ export function fixViteErrorMessage(_err: unknown, server: ViteDevServer) {
 		err.message = 'Astro.glob() and import.meta.glob() can only accept string literals.';
 	}
 	return err;
+}
+
+const incompatiblePackages = {
+	'react-spectrum': `@adobe/react-spectrum is not compatible with Vite's server-side rendering mode at the moment. You can still use React Spectrum from the client. Create an island React component and use the client:only directive. From there you can use React Spectrum.`,
+};
+const incompatPackageExp = new RegExp(`(${Object.keys(incompatiblePackages).join('|')})`);
+
+function generateHint(err: ErrorWithMetadata): string | undefined {
+	if (/Unknown file extension \"\.(jsx|vue|svelte|astro|css)\" for /.test(err.message)) {
+		return 'You likely need to add this package to `vite.ssr.noExternal` in your astro config file.';
+	} else {
+		const res = incompatPackageExp.exec(err.stack);
+		if (res) {
+			const key = res[0] as keyof typeof incompatiblePackages;
+			return incompatiblePackages[key];
+		}
+	}
+	return undefined;
 }
 
 /**
@@ -70,9 +89,11 @@ export function collectErrorMetadata(e: any): ErrorWithMetadata {
 		if (pluginName) {
 			err.plugin = pluginName;
 		}
+		err.hint = generateHint(err);
 		return err;
 	}
 
 	// Generic error (probably from Vite, and already formatted)
+	e.hint = generateHint(e);
 	return e;
 }