+++ packages/astro/test/errors.test.js
@@ -1,9 +1,11 @@
 import { isWindows, loadFixture } from './test-utils.js';
 import { expect } from 'chai';
+import * as cheerio from 'cheerio';
 
 describe('Error display', () => {
 	if (isWindows) return;
 
+	/** @type {import('./test-utils').Fixture} */
 	let fixture;
 	let devServer;
 
@@ -28,6 +30,34 @@ describe('Error display', () => {
 			const res = await fixture.fetch('/astro-syntax-error');
 			await devServer.stop();
 			expect(res.status).to.equal(500, `Successfully responded with 500 Error for invalid file`);
+		});
+	});
+
+	describe('Framework components', function () {
+		let devServer;
+
+		before(async () => {
+			devServer = await fixture.startDevServer();
+		});
+
+		after(async () => {
+			await devServer.stop();
+		});
+
+		it('Errors recover when fixed', async () => {
+			let html = await fixture.fetch('/svelte-syntax-error').then((res) => res.text());
+
+			// 1. Verify an error message is being shown.
+			let $ = cheerio.load(html);
+			expect($('.statusMessage').text()).to.equal('Internal Error');
+
+			// 2. Edit the file, fixing the error
+			await fixture.editFile('./src/components/SvelteSyntaxError.svelte', `<h1>No mismatch</h1>`);
+
+			// 3. Verify that the file is fixed.
+			html = await fixture.fetch('/svelte-syntax-error').then((res) => res.text());
+			$ = cheerio.load(html);
+			expect($('h1').text()).to.equal('No mismatch');
 		});
 	});
 });